Change to i32

+++ src/objects/spread.rs
@@ -2,8 +2,8 @@ use serde::{Deserialize, Serialize};
 
 #[derive(Debug, Clone, Serialize, Deserialize)]
 pub struct Spread {
-	pub home_id: i64,
-	pub away_id: i64,
-	pub home_spread: f64,
-	pub notes: Option<String>
+    pub home_id: i32,
+    pub away_id: i32,
+    pub home_spread: f64,
+    pub notes: Option<String>,
 }
